explore search + updated my guides header
ui changes for explore search. changed some data referencing in search bar but still have to fix it once backend data is passed in. updated header of my guides to be consistent with explore page. Updated stack: screens are now in app.js so they do not show tab bar